refactor: created test util for common actions
